% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fib_cpp_1}
\alias{fib_cpp_1}
\title{A function of Fibonacci sequence using Rcpp}
\usage{
fib_cpp_1(n)
}
\arguments{
\item{n}{the nth term of the Fibonacci sequence}
}
\value{
a int number
}
\description{
A function to calculate the nth term of the Fibonacci sequence
}
\examples{
\dontrun{
fib1 <- fib_cpp_1(10)
}
}
