% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naivekernelplot.R
\name{naivekernelplot}
\alias{naivekernelplot}
\title{A plot function of naive kernel estimation}
\usage{
naivekernelplot(N, h)
}
\arguments{
\item{N}{the number of random number}

\item{h}{the bandwidth of the naive kernel estimation}
}
\value{
a plot of a naive kernel estimation function
}
\description{
By the function, we can estimate a mixed normal distribution by the naive kernel estimation and draw the plot.
}
\examples{
\dontrun{
nk <- naivekernelplot(1000,0.1)
par(mfrow=c(1,1));
nk
}
}
