% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redata.R
\name{redata}
\alias{redata}
\title{A function to generate random number}
\usage{
redata(k, l, fl, e)
}
\arguments{
\item{k}{the number of random number}

\item{l}{the dimensions of linear random number changed by time}

\item{fl}{the dimensions of log random number changed by time}

\item{e}{the dimensions of exponential random number changed by time}
}
\value{
a dataframe of random numbers
}
\description{
A function to generate random number for my research. This function can control the number and the dimensions of of the random numbers. And it can generate three kinds of random numbers.
}
\examples{
\dontrun{
redata(10,1,1,1)
}
}
