\name{StatComp20048-package}
\alias{StatComp20048-package}
\alias{StatComp20048}
\docType{package}
\title{
  StatComp20048
}
\description{
  R package that include at least two functions from my research project and all functions for my homework
}
\details{
  R package that include at least two functions from my research project and all functions for my homework
}
\author{
Guo Bangwei

Maintainer: Guo Bangwei <gbw0011@mail.ustc.edu.cn>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
