% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum3}
\alias{sum3}
\title{A sum function using Rcpp}
\usage{
sum3(x)
}
\arguments{
\item{x}{the numeric vector}
}
\value{
sum
}
\description{
A function to calculate the sum of a vector
}
\examples{
\dontrun{
rnC <- sum3(1:10)
}
}
